functions/building_chat/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.39.11
botocore==1.39.11
certifi==2025.7.14
charset-normalizer==3.4.2
ecdsa==0.19.1
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jmespath==1.0.1
numpy==1.26.4
portalocker==3.2.0
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-jose==3.5.0
qdrant-client==1.15.0
requests==2.32.4
rsa==4.9.1
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
